use AptUrlRedirector instead of plain apt:// URLs in README
README.md isn't displayed on the Jekyll-generated website (or if it is, I'm not aware where), so the primary way to view it is on the GitHub repository's page. GitHub MarkDown doesn't support links with apt:// URLs. Thus, use redirection HTTP URLs provided by Ubuntu as documented on https://wiki.ubuntu.com/AptUrlRedirector